+++ trunk/xmipp/libraries/reconstruction/nma_alignment.h
@@ -166,7 +166,7 @@ public:
         An exception is thrown if any of the files is not found*/
     virtual void preProcess();
     /** Assign NMA and Alignment parameters to an image */
-    virtual void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId);
+    virtual void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId);
     virtual void writeImageParameters(const FileName &fnImg);
 };
